rename service/data_storage to service/dataset module
