+++ service/dataset/upload_dir.go
@@ -1,4 +1,4 @@
-package data_storage
+package dataset
 
 import (
 	"context"
